Rename NoteActivity -> AddNoteActivity
