+++ presentation/src/main/AndroidManifest.xml
@@ -24,7 +24,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".feature.note.NoteActivity">
+        <activity android:name=".feature.addnote.AddNoteActivity">
         </activity>
         <activity android:name=".feature.notedetails.DetailsActivity">
         </activity>
