+++ presentation/src/main/java/by/grodno/vasili/presentation/feature/addnote/AddNoteActivityScope.java
@@ -1,4 +1,4 @@
-package by.grodno.vasili.presentation.feature.note;
+package by.grodno.vasili.presentation.feature.addnote;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -7,5 +7,5 @@ import javax.inject.Scope;
 
 @Scope
 @Retention(RetentionPolicy.RUNTIME)
-public @interface NoteActivityScope {
+public @interface AddNoteActivityScope {
 }
